requirements.txt
langgraph>=0.0.40
langchain-core>=0.1.15

gspread>=5.12.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0

pandas>=2.0.0
python-dotenv>=1.0.0

